Add files via upload
These are the datasets utlilised for my data analysis. The cardiovascular.csv file is a small data set with less than 1,000 rows while the cardio_train.csv is a larger data set with more than 100,000 rows. With the larger data set, we decided to utilize GPU for the analytics and predictions.